backend/requirements-minimal.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=1.7.4,<2.1.0
